requirements.txt
Django==1.5.4
South==0.8.4
amqp==1.0.6
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.19
celery==3.0.13
dj-database-url==0.2.1
django-celery==3.0.11
django-storages==1.1.8
psycopg2==2.5.1
python-dateutil==1.5
requests==1.2.0
gunicorn==0.16.1
simplejson==3.1.3
six==1.2.0
wsgiref==0.1.2
